many warnings for GCC removed (snide/prissi)
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@2010 8aca7d54-2c30-db11-9de9-000461428c89

+++ besch/haus_besch.cc
@@ -16,7 +16,7 @@
  *  Beschreibung:
  *      Rechnet aus dem Index das Layout aus, zu dem diese Tile gehört.
  */
-int haus_tile_besch_t::gib_layout() const
+uint8 haus_tile_besch_t::gib_layout() const
 {
     koord groesse = gib_besch()->gib_groesse();
