+++ gui/fahrplan_gui.cc
@@ -149,7 +149,7 @@ void fahrplan_gui_stats_t::zeichnen(koord offset)
 		cbuffer_t buf(512);
 		image_id const arrow_right_normal = skinverwaltung_t::window_skin->gib_bild(10)->gib_nummer();
 
-		for (uint32 i = 0; i < fpl->maxi(); i++) {
+		for (int i = 0; i < fpl->maxi(); i++) {
 
 			buf.clear();
 			buf.printf( "%i) ", i+1 );
@@ -243,7 +243,7 @@ fahrplan_gui_t::fahrplan_gui_t(fahrplan_t* fpl_, spieler_t* sp_, convoihandle_t 
 	bt_wait_prev.add_listener(this);
 	add_komponente(&bt_wait_prev);
 
-	if((unsigned)fpl->aktuell>=fpl->maxi()  ||  fpl->eintrag[fpl->aktuell].waiting_time_shift==0) {
+	if(fpl->aktuell>=fpl->maxi()  ||  fpl->eintrag[fpl->aktuell].waiting_time_shift==0) {
 		strcpy( str_parts_month, translator::translate("off") );
 	}
 	else {
@@ -269,7 +269,7 @@ fahrplan_gui_t::fahrplan_gui_t(fahrplan_t* fpl_, spieler_t* sp_, convoihandle_t 
 	bt_prev.add_listener(this);
 	add_komponente(&bt_prev);
 
-	sprintf( str_ladegrad, "%d%%", (unsigned)fpl->aktuell<fpl->maxi() ? fpl->eintrag[fpl->aktuell].ladegrad : 0 );
+	sprintf( str_ladegrad, "%d%%", fpl->aktuell<fpl->maxi() ? fpl->eintrag[fpl->aktuell].ladegrad : 0 );
 	lb_loadlevel.set_text_pointer( str_ladegrad );
 	lb_loadlevel.pos = koord( BUTTON_WIDTH*2-20, ypos+3 );
 	add_komponente(&lb_loadlevel);
